# Make EnforceSorting global sort aware, fix sort mis-optimizations involving unions, support parallel sort + merge transformations