--- a/datafusion/core/src/execution/context.rs
+++ b/datafusion/core/src/execution/context.rs
@@ -1221,6 +1221,12 @@ impl SessionConfig {
         self.options.optimizer.repartition_windows
     }
 
+    /// Are Sorts distributed to multiple cores, If there is already
+    /// multiple cores in the plan?
+    pub fn parallelize_sorts(&self) -> bool {
+        self.options.optimizer.parallelize_sorts
+    }
+
     /// Are statistics collected during execution?
     pub fn collect_statistics(&self) -> bool {
         self.options.execution.collect_statistics
@@ -1276,6 +1282,12 @@ impl SessionConfig {
     /// Enables or disables the use of repartitioning for window functions to improve parallelism
     pub fn with_repartition_windows(mut self, enabled: bool) -> Self {
         self.options.optimizer.repartition_windows = enabled;
+        self
+    }
+
+    /// Enables or disables the use of Sort parallelization
+    pub fn with_parallelize_sorts(mut self, enabled: bool) -> Self {
+        self.options.optimizer.parallelize_sorts = enabled;
         self
     }
 
@@ -1508,10 +1520,8 @@ impl SessionState {
             // The EnforceSorting rule is for adding essential local sorting to satisfy the required
             // ordering. Please make sure that the whole plan tree is determined before this rule.
             // Note that one should always run this rule after running the EnforceDistribution rule
-            // as the latter may break local sorting requirements. The rule takes a boolean flag
-            // indicating whether we elect to transform CoalescePartitionsExec + SortExec cascades
-            // into SortExec + SortPreservingMergeExec cascades, enabling parallel sorting.
-            Arc::new(EnforceSorting::new(true)),
+            // as the latter may break local sorting requirements.
+            Arc::new(EnforceSorting::new()),
             // The CoalesceBatches rule will not influence the distribution and ordering of the
             // whole plan tree. Therefore, to avoid influencing other rules, it should run last.
             Arc::new(CoalesceBatches::new()),