… datafusion/core/tests/sql/joins.rs
… datafusion/core/tests/sql/window.rs
… datafusion/core/tests/sqllogictests/test_files/information_schema.slt
… docs/source/user-guide/configs.md