@alamb 2023-02-06
BTW I checked this code against IOx's tests -- https://github.com/influxdata/influxdb_iox/pull/6870 -- it looks good to me