@ozankabak 2023-02-06
Thanks for reviewing! We will address your reviews and finalize this tomorrow. FYI, @mingmwang was kind enough to go through this PR in our fork, and has given us some examples to test the changes on.

After this merges with the new functionality, he plans to propose a top-down refactoring that simplifies the code without losing the functionality -- which I am super enthusiastic to see.