@alamb 2023-02-06
> After this merges with the new functionality, he plans to propose a top-down refactoring that simplifies the code without losing the functionality -- which I am super enthusiastic to see.

I agree this sounds great

For anyone else following along, I think the discussion @ozankabak  is referring to is on https://github.com/synnada-ai/arrow-datafusion/pull/43