@mingmwang 2023-02-07
@alamb @ozankabak @mustafasrepo 

Regarding the global sort replaced to a parallel version(SortPreservingMergeExec + Local Sort) optimization, I think there is already a rule `GlobalSortSelection` for the exact purpose. I think we should not let  the Sort Enforcement rule to handle this again.