@mingmwang 2023-02-07
@mustafasrepo @ozankabak 
Regarding the rule applying ordering, since DataFusion optimization framework is still a traditional heuristic style framework, the rule applying orders always matter, we can not assume one rule can work independently without the others.  

Specifically , `EnforceDistribution` rule is responsible for handling the global distribution requirements.
And `EnforceSorting` rule is responsible for handling the local sort requirements. It's also responsible for removing
unnecessary global sort and local sort.   The global distribution requirements need to be handled first, after that we can handle the local sort(inner-partition) requirements.

Global properties vs Local properties
http://www.cs.albany.edu/~jhh/courses/readings/zhou10.pdf